+++ server/src/main/java/org/elasticsearch/common/io/stream/RecyclerBytesStreamOutput.java
@@ -13,11 +13,9 @@ import org.apache.lucene.util.BytesRef;
 import org.elasticsearch.common.bytes.BytesArray;
 import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.bytes.CompositeBytesReference;
-import org.elasticsearch.common.compress.CompressorFactory;
 import org.elasticsearch.common.recycler.Recycler;
 import org.elasticsearch.core.Releasable;
 import org.elasticsearch.core.Releasables;
-import org.elasticsearch.core.Streams;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
@@ -157,7 +155,7 @@ public class RecyclerBytesStreamOutput extends BytesStream implements Releasable
     }
 
     @Override
-    public void writeWithSizePrefix(Writeable writeable) throws IOException {
+    public void legacyWriteWithSizePrefix(Writeable writeable) throws IOException {
         // TODO: do this without copying the bytes from tmp by calling writeBytes and just use the pages in tmp directly through
         // manipulation of the offsets on the pages after writing to tmp. This will require adjustments to the places in this class
         // that make assumptions about the page size
@@ -177,20 +175,6 @@ public class RecyclerBytesStreamOutput extends BytesStream implements Releasable
                 tmpPage++;
             }
         }
-    }
-
-    @Override
-    public void writeWithSizePrefix2(Writeable writeable) throws IOException {
-        long pos = position();
-        skip(Integer.BYTES);
-        try (var out = new OutputStreamStreamOutput(CompressorFactory.COMPRESSOR.threadLocalOutputStream(Streams.noCloseStream(this)))) {
-            out.setTransportVersion(getTransportVersion());
-            writeable.writeTo(out);
-        }
-        long newPos = position();
-        seek(pos);
-        writeInt(Math.toIntExact(newPos - pos - Integer.BYTES));
-        seek(newPos);
     }
 
     // overridden with some code duplication the same way other write methods in this class are overridden to bypass StreamOutput's
@@ -230,6 +214,7 @@ public class RecyclerBytesStreamOutput extends BytesStream implements Releasable
         // nothing to do
     }
 
+    @Override
     public void seek(long position) {
         ensureCapacityFromPosition(position);
         int offsetInPage = (int) (position % pageSize);
